Refactor comment validation
